Made Changes to Level 6 and 7.
Added a wall in level 6 to make the player go a longer way to the exit.
Added a all into level 7 to show the player that they cannot move left
or right at the edge of the screen. Also fixed the collision bug at the
top of the spikes by making the panel above them one surface instead of
three.