New levels and mechanics
I have created a script to control the movement of spikes and other
mechanics. You have the options to; ping pong, move on x and y, pause at
the start, pause before it moves back to it's original position - the
offset and to set which way you would like to move it first. There is
also a prefab that shows how it works.

I have created a script to enable objects to follow a track. You can
create target points that you would like the object to move to. This
allows for more complex paths to be planned easily with a visual
representation. This script will still need to rotate to the next
targets position. Also this should have a toggle option in the inspector
panel.

I have added six new levels that may need changing. They range in
difficulty and will have to be put in order to build the player up
gradually. All have new mechanics in that are being used in a different
way to what the player has seen before. These are spikes but the code
base allows for more mechanics to be added easily.

I have tidied up some of the other mechanics scripts to make them more
readable.