Added four levels & added new a parentscript
New prefabs have been added for the movable base floor. Also the long
and unneeded parent to object script have been simplified.

Four new levels with existing mechanics. The movementTransformScript
will have to be changed to velocity based movement to get rid of the
small move while the player is on moving platforms.